% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{polygon}
\alias{polygon}
\title{GET data from Polygon.io API}
\usage{
polygon(path, args = NULL, ...)
}
\arguments{
\item{path}{API path by a character scalar}

\item{args}{Query arguments by named list or named vector}

\item{...}{Addtional arguments for httr::GET()}
}
\value{
API response parsed by jsonlite::fromJSON()
}
\description{
GET data from Polygon.io API
}
