% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{polygon_reference}
\alias{polygon_reference}
\title{Retrieve reference data}
\usage{
polygon_reference(type = "markets", args = NULL, ...)
}
\arguments{
\item{type}{Reference type}

\item{args}{Query arguments}

\item{...}{Addtional arguments for httr::GET()}
}
\value{
list(status, results)
}
\description{
Retrieve reference data
}
