% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpolygon.R
\name{rpolygon}
\alias{rpolygon}
\title{GET data from Polygon.io API}
\usage{
rpolygon(endpoint, args = NULL, ...)
}
\arguments{
\item{endpoint}{API endpoint by a character scalar}

\item{args}{Query arguments by named list or named vector}

\item{...}{Addtional arguments for httr::GET()}
}
\value{
API response parsed by jsonlite::fromJSON()
}
\description{
GET data from Polygon.io API
}
