% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{rpolygon_historic}
\alias{rpolygon_historic}
\title{GET daily trades/quote}
\usage{
rpolygon_historic(symbol, date, type = "trades")
}
\arguments{
\item{symbol}{Symbol}

\item{date}{Date}

\item{type}{trades or quotes}
}
\value{
data.frame
}
\description{
GET daily trades/quote
}
